+++ src/shopify-auth/controllers/authHandler.ts
@@ -1,6 +1,7 @@
 import Shopify from '@shopify/shopify-api';
 import { Request, Response } from 'express';
 import { ShopifyAuthModuleOptions } from '../interfaces';
+import { joinUrl } from '../utils/join-url.util';
 
 export async function authHandler(
   req: Request,
@@ -11,10 +12,8 @@ export async function authHandler(
   globalPrefix = '',
 ) {
   const { basePath } = options;
-  const redirectUrl = `${globalPrefix}/${basePath}/callback`.replace(
-    /\/\//g,
-    '/',
-  );
+
+  const redirectUrl = joinUrl(globalPrefix, basePath, 'callback');
 
   const oauthUrl = await Shopify.Auth.beginAuth(
     req,